feat: Nova 4 support
